Fixed Dividend Issue
Fixed an issue where if an equation had a dividend but the ticker didn't
have a dividend, the program would not execute. I split it up using an
if and an else statement. If the program didn't have a dividend then the
program would execute all equations inside of the if statement.